+++ src/EFCore.Ydb/CHANGELOG.md
@@ -1,3 +1,5 @@
+## v0.2.0
+
 - Fixed bug: Missing/removed __EFMigrationsHistory caused INSERT errors (LockMigration); migrations now handle table recreation and locking properly ([#544](https://github.com/ydb-platform/ydb-dotnet-sdk/issues/544)).
 - Added support for new YDB data types - `Date32`, `Datetime64`, `Timestamp64` and `Interval64`.
 - Fixed bug: Missing `DEFAULT` clause generation in `YdbMigrationsSqlGenerator` (handles defaultValue/defaultValueSql) ([#552](https://github.com/ydb-platform/ydb-dotnet-sdk/issues/552)).